pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openbrain-gateway"
version = "0.1.0"
description = "Canonical memory gateway for OpenClaw + Open Brain + Brian Madden federation"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.35",
  "psycopg[binary]>=3.2.1",
  "pgvector>=0.3.5",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.2",
  "httpx>=0.27.2",
  "typer>=0.12.5",
  "python-slugify>=8.0.4",
  "sentence-transformers>=3.2.0",
  "numpy>=2.1.0",
  "python-frontmatter>=1.1.0"
]

[project.scripts]
openbrainctl = "openbrain.cli:app"

[tool.setuptools.packages.find]
include = ["openbrain*"]

